Remove all_notebooks option and sort out conftest
all_notebooks option doesn't seem useful.
